# backend/auth-service/src/main/java/com/company/platform/auth/user/dto/request/UpdateUserRequest.java
package com.company.platform.auth.user.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter @Setter @NoArgsConstructor
public class UpdateUserRequest {
    @Email
    private String email;

    @Size(max = 200)
    private String fullName;
}