# docs/api/auth-openapi-v1.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  description: Identity, authentication, and RBAC management API for Enterprise CRM Platform.
  version: 1.0.0
  contact:
    name: Platform Engineering Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development (via API Gateway)
  - url: http://localhost:8081/api/v1
    description: Auth service direct access
...
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate username or email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Users]
      summary: List users with pagination
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Users]
      summary: Update user profile
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/status:
    patch:
      tags: [Users]
      summary: Update user status
      operationId: updateUserStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/roles:
    post:
      tags: [Users]
      summary: Assign roles to user
      operationId: assignRoles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest'
      responses:
        '200':
          description: Roles assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignRolesResponse'
        '403':
          description: Privilege escalation denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/roles/{roleId}:
    delete:
      tags: [Users]
      summary: Remove role from user
      operationId: removeRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role removed
        '404':
          description: User or role mapping not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: createdAt,desc
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
        enum: [ACTIVE, INACTIVE, LOCKED]
  schemas:
...
    CreateUserRequest:
      type: object
      required: [username, email, password, fullName]
      properties:
        username:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
        fullName:
          type: string
          maxLength: 200
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          default: ACTIVE
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
          maxLength: 200
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    UserDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        status:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleSummary'
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    RoleSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        roleCode:
          type: string
        roleName:
          type: string
    StatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
    AssignRolesRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
    AssignRolesResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleSummary'
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: string
          format: uuid
...
    SearchSort:
      type: object
      required: [field, direction]
      properties:
        field:
          type: string
        direction:
          type: string
          enum: [ASC, DESC]
    PagedUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SearchSort'
...
    ErrorResponse:
      type: object
      required: [code, message, traceId, timestamp]
      properties:
        code:
          type: string
          description: Machine-readable error code (AUTH_ prefix)
        message:
          type: string
          description: Human-readable error description
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          description: Field-level validation errors (for 400/422)
        traceId:
          type: string
          description: Distributed trace correlation ID
        timestamp:
          type: string
          format: date-time
          description: Error timestamp in UTC
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string